Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15502 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/ADSrc/asynNDArrayDriver.h
@@ -161,5 +161,5 @@ protected:
                                           *  attributes */
 };
 
-#define NUM_NDARRAY_PARAMS (&LAST_NDARRAY_PARAM - &FIRST_NDARRAY_PARAM + 1)
+#define NUM_NDARRAY_PARAMS ((int)(&LAST_NDARRAY_PARAM - &FIRST_NDARRAY_PARAM + 1))
 #endif
